refactor: split version managers
